Jichen pre release (#24)
* [network] update modularized pinet2

* [doc] update pinet2
