--- inputs/README.md
@@ -1,14 +1,0 @@
-# Inputs for potentials
-
-* pinet
-    - nofrc: energy-only training
-* pinet2
-    - default: weighted = False, enegy & force used in training
-    - weighted: pinet2 weighted = True (default False)
-    - frc: force-only training
-    - nofrc: energy-only training
-
-* units:
-    * energy: kcal/mol
-    * force: kcal/mol/Angstrom
-    * distance: Angstrom
