--- inputs/frc/pinet-pot.yml
@@ -1,31 +1,0 @@
-model:
-  name: potential_model
-  params:
-    use_force: true
-    e_loss_multiplier: 0.0
-    f_loss_multiplier: 1.0
-    e_scale: 1.0
-    e_unit: 1.0
-    use_e_per_atom: false
-    log_e_per_atom: true
-network:
-  name: PiNet
-  params:
-    depth: 5
-    rc: 4.5
-    n_basis: 10
-    basis_type: gaussian
-    pi_nodes: [64]
-    pp_nodes: [64,64,64,64]
-    ii_nodes: [64,64,64,64]
-    out_nodes: [64]
-optimizer:
-  class_name: Adam
-  config:
-    global_clipnorm: 0.01
-    learning_rate:
-      class_name: ExponentialDecay
-      config:
-        decay_rate: 0.994
-        decay_steps: 10000
-        initial_learning_rate: 0.0001
