+++ nextflow.config
@@ -31,11 +31,13 @@ profiles {
   }
   alvis {
     singularity.enabled = false
-    process.scratch = true
-    process.time = '3d'
-    process.clusterOptions = '--gres=gpu:T4:1'
-    process.beforeScript = 'source $HOME/pinn-tf26/bin/activate'
-    process.module = 'TensorFlow/2.6.0-foss-2021a-CUDA-11.3.1'
+    process.scratch = false
+    process.time = '6d'
+    process.clusterOptions = '--gres=gpu:V100:1'
+    process.beforeScript = """
+    module load TensorFlow/2.11.0-foss-2022a-CUDA-11.7.0.lua
+    source /mimer/NOBACKUP/groups/teoroo_1/jichen/pinn-tf211/bin/activate
+    """
     executor{
       name = 'slurm'
       queueSize = 20
